Berechnung von Co2 Fix
Dauerfahrt bug, Wochenübersicht bug

+++ app/src/main/java/com/example/prototype_footprinthero/view/MainScreen.kt
@@ -55,7 +55,7 @@ fun MainScreen(viewModel: MainViewModel) {
             Row(Modifier.padding(16.dp), verticalAlignment = Alignment.CenterVertically) {
                 Box(modifier = Modifier.weight(1f)) {
                     TransportationDurationView(
-                        duration = viewModel.duration.value.toFloat(),
+                        duration = viewModel.duration.value,
                         onDurationChanged = viewModel::onDurationChanged
                     )
                 }
